валидация формы перед и после отправки
