+++ css/style.css
@@ -69,6 +69,7 @@ fieldset {
 
 .map {
   position: relative;
+  z-index: 1; /*чтобы карта не перекрывала информационное сообщение при отправке*/
   display: -webkit-box;
   display: -ms-flexbox;
   display: flex;
