+++ js/form-validation.js
@@ -1,76 +1,40 @@
-import {adForm} from "./map-baloon.js";
-import {rentPrice} from "./auto-change-input.js";
+import {isEscEvent} from "./util.js";
+import {adForm, InputAddress, coordCenter, mainMarker, mapFiltersForm} from "./activate-page.js";
 
-const formTitle = adForm.querySelector("#title");
-const roomNumber = adForm.querySelector("#room_number");
-const capacity = adForm.querySelector("#capacity");
+// eslint-disable-next-line no-undef
+const latLng = L.latLng(coordCenter);
 
-const MIN_TITLE_LENGTH = 30;
-const MAX_TITLE_LENGTH = 100;
-const MAX_PRICE_LENGTH = 1000000;
+const successMassage = document.querySelector("#success")
+    .content
+    .querySelector(".success");
 
+const successFragment = document.createDocumentFragment();
+const successClone = successMassage.cloneNode(true);
 
-formTitle.addEventListener ("input", () => {
-    const valueLength = formTitle.value.length;
-    if (valueLength < MIN_TITLE_LENGTH) {
-        formTitle.setCustomValidity("Минимальная длина — 30 символов, " + "ещё " + (MIN_TITLE_LENGTH - valueLength));
-    }
-    else if (valueLength > MAX_TITLE_LENGTH) {
-        formTitle.setCustomValidity("Максимальная длина — 100 символов");
-    } else if (formTitle.validity.valueMissing) {
-        formTitle.setCustomValidity("Обязательное поле");
-    } else {
-        formTitle.setCustomValidity(""); //сброс сообщения, если ошибок больше нет
-    }
-    formTitle.reportValidity();//отслеживать каждый ввод с клавиатуры
+successFragment.appendChild(successClone);
+
+const failMassage = document.querySelector("#error")
+    .content
+    .querySelector(".error");
+
+const failFragment = document.createDocumentFragment();
+const failClone = failMassage.cloneNode(true);
+
+failFragment.appendChild(failClone);
+
+const btnError = failClone.querySelector(".error__button");
+const btnReset = adForm.querySelector(".ad-form__reset");
+const coordMarker = mainMarker.getLatLng();
+const tagBody = document.querySelector("body");
+
+btnReset.addEventListener ("click", (evt) => {
+    evt.preventDefault();
+    adForm.reset();
+    mapFiltersForm.reset();
+    mainMarker.setLatLng(latLng);
+    InputAddress.value =  parseFloat(coordMarker.lat.toFixed(5)) + ", " + parseFloat(coordMarker.lng.toFixed(5));
+
 });
-
-rentPrice.addEventListener ("input", () => {
-    const valuePriceNumber = rentPrice.value;
-    if (valuePriceNumber > MAX_PRICE_LENGTH) {
-        rentPrice.setCustomValidity("Максимальное значение — 1 000 000");
-    } else if (rentPrice.validity.valueMissing) {
-        rentPrice.setCustomValidity("Обязательное поле");
-    } else {
-        rentPrice.setCustomValidity(""); //сброс сообщения, если ошибок больше нет
-    }
-    rentPrice.reportValidity();//отслеживать каждый ввод с клавиатуры
-});
-
-const checkRoomsNumber = () => {
-    if(roomNumber.value === "1") {
-        if(capacity.value !==roomNumber.value) {
-            capacity.setCustomValidity("Максимальное количество гостей - 1 гость");
-        }else {
-            capacity.setCustomValidity(""); //сброс сообщения, если ошибок больше нет
-        }
-    }
-    if(roomNumber.value === "2") {
-        if(capacity.value==="3"|| capacity.value==="0") {
-            capacity.setCustomValidity("количество гостей - 2 гостя или 1 гость");
-        }else {
-            capacity.setCustomValidity(""); //сброс сообщения, если ошибок больше нет
-        }
-    }
-    if(roomNumber.value === "3") {
-        if(capacity.value==="0") {
-            capacity.setCustomValidity("Выберите количество гостей: 1,2 или 3 гостя");
-        }else {
-            capacity.setCustomValidity(""); //сброс сообщения, если ошибок больше нет
-        }
-    }
-    if(roomNumber.value === "100") {
-        if(capacity.value!=="0") {
-            capacity.setCustomValidity("Не для гостей");
-        }else {
-            capacity.setCustomValidity(""); //сброс сообщения, если ошибок больше нет
-        }
-    }
-    capacity.reportValidity();
-};
-checkRoomsNumber();
-
-capacity.addEventListener ("change", () => checkRoomsNumber());
 
 adForm.addEventListener("submit", (evt) => {
     evt.preventDefault();
@@ -82,5 +46,40 @@ adForm.addEventListener("submit", (evt) => {
             method: "POST",
             body: formData,
         },
-    );
+    )   .then((response) => response.json())
+        .then(() => {
+            tagBody.appendChild(successFragment);
+            successClone.classList.remove("hidden");
+            document.addEventListener("keydown", (evt) => {
+                if (isEscEvent(evt)) {
+                    evt.preventDefault();
+                    successClone.classList.add("hidden");
+                }
+            });
+            document.addEventListener("click", () => {
+                successClone.classList.add("hidden");
+            });
+            evt.target.reset();
+            mapFiltersForm.reset();
+            mainMarker.setLatLng(latLng);
+            InputAddress.value =  parseFloat(coordMarker.lat.toFixed(5)) + ", " + parseFloat(coordMarker.lng.toFixed(5));
+        })
+        .catch(() => {
+            tagBody.appendChild(failFragment);
+            failClone.classList.remove("hidden");
+            document.addEventListener("keydown", (evt) => {
+                if (isEscEvent(evt)) {
+                    evt.preventDefault();
+                    failClone.classList.add("hidden");
+                }
+            });
+            document.addEventListener("click", () => {
+                failClone.classList.add("hidden");
+            });
+            btnError.addEventListener("click", () => {
+                failClone.classList.add("hidden");
+            });
+        });
 });
+
+
