+++ js/main.js
@@ -1,6 +1,5 @@
 import "./util.js";
 import "./data.js";
-import "./layout-creator.js";
 import "./auto-change-input.js";
 import "./map-baloon.js";
 import "./form-validation.js";
