+++ js/util.js
@@ -43,4 +43,41 @@ const getRandomArray = (someArray) => {
     return newMass;
 };
 
-export{getRandomArrayElement, getRandomArray};
+const showAlert = (message) => {
+    const alertContainer = document.createElement("div");
+    alertContainer.style.zIndex = 100;
+    alertContainer.style.position = "absolute";
+    alertContainer.style.left = 0;
+    alertContainer.style.top = 0;
+    alertContainer.style.right = 0;
+    alertContainer.style.padding = "10px 3px";
+    alertContainer.style.fontSize = "30px";
+    alertContainer.style.textAlign = "center";
+    alertContainer.style.backgroundColor = "#D2691E";
+
+    alertContainer.textContent = message;
+
+    document.body.append(alertContainer);
+};
+
+const isEscEvent = (evt) => {
+    return evt.key === "Escape" || evt.key === "Esc";
+};
+
+const makeDisable = (unit) => {
+    if (Array.isArray(unit)) {
+        unit.forEach((element)=> {
+            element.disabled = "disabled";
+        });
+    }  unit.disabled = "disabled";
+};
+
+const removeDisable = (unit) => {
+    if (Array.isArray(unit)){
+        unit.forEach((element)=>{
+            element.disabled = "";
+        });
+    }  unit.disabled = "";
+};
+
+export{getRandomArrayElement, getRandomArray, showAlert, isEscEvent, makeDisable, removeDisable};
